Plot API now has Independent Plots (#415)
* fixed docs issue #338

* maybe fixed github issue #352

* fixed issue #352 for int and float at least

* typo in the error msg of get_int_visitor

* removed dead code

* made error message more informative by including value of invalid string

* added unit test for type conversion

* added space in the error message, althought I think it gets caught in unity_sarray and hence gets silenced by the error message in unity_sarray on line 1336/1340?

* addressed yucheng's concerns

* not using \0 check anymore: checking whether data() + length() == end_ptr or not

* added code for all independent plots except item_frequency

* python code to complete all the code for independent plots

* code complete for independent plots

* Code complete, built, and tested all independent plots -- Independent plots work

* gl_sarray(self)
