+++ src/unity/python/turicreate/visualization/__init__.py
@@ -9,3 +9,10 @@ from __future__ import absolute_import as _
 from .show import show
 from .plot import Plot, set_target
 from .show import _get_client_app_path
+from .show import scatter
+from .show import categorical_heatmap
+from .show import heatmap
+from .show import box_plot
+from .show import columnwise_summary
+from .show import histogram
+from .show import item_frequency
